review fix

+++ source/MRCommonPlugins/ViewerButtons/MRViewerSettingsPlugin.cpp
@@ -98,7 +98,7 @@ void ViewerSettingsPlugin::drawDialog( float menuScaling, ImGuiContext* )
 
 void ViewerSettingsPlugin::addComboSettings( const TabType tab, std::shared_ptr<ComboSettings> settings )
 {
-    comboSettings_[tab].push_back( settings );
+    comboSettings_[size_t( tab )].push_back( settings );
 }
 
 bool ViewerSettingsPlugin::onEnable_()
@@ -558,7 +558,7 @@ void ViewerSettingsPlugin::drawTouchpadSettings_()
 
 void ViewerSettingsPlugin::drawCustomSettinds_( TabType tabType )
 {
-    for ( auto& settings : comboSettings_[tabType] )
+    for ( auto& settings : comboSettings_[size_t( tabType )] )
     {
         settings->draw();
     }
